move test from core to optimizer crate
