+++ datafusion/physical-optimizer/Cargo.toml
@@ -33,8 +33,14 @@ workspace = true
 
 [dependencies]
 arrow-schema = { workspace = true }
+arrow = { workspace = true }
 datafusion-common = { workspace = true, default-features = true }
 datafusion-execution = { workspace = true }
+datafusion-expr-common = { workspace = true, default-features = true }
 datafusion-physical-expr = { workspace = true }
 datafusion-physical-plan = { workspace = true }
 itertools = { workspace = true }
+
+[dev-dependencies]
+datafusion-functions-aggregate = { workspace = true }
+tokio = { workspace = true }
